requirements.txt
fire==0.2.1
mysql-connector-python==8.0.18
postgres-create-tables==0.1
protobuf==3.10.0
six==1.12.0
termcolor==1.1.0
